# api-tests/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Noah Explorer api",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/v1/validators/{pubKey}" : {
      "get" : {
        "tags" : [ "Validators" ],
        "summary" : "Get validator detail by public key",
        "operationId" : "getValidator",
        "parameters" : [ {
          "name" : "pubKey",
          "in" : "path",
          "description" : "Validator public key",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Validator detail",
            "content" : {
              "application-json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidatorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Validator not found",
            "content" : {
              "application-json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_404_1"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/validators/{pubKey}/transactions" : {
      "get" : {
        "tags" : [ "Validators" ],
        "summary" : "Get list of transactions by validator",
        "operationId" : "getValidatorTransactions",
        "parameters" : [ {
          "name" : "pubKey",
          "in" : "path",
          "description" : "Validator public key",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page number",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "startblock",
          "in" : "query",
          "description" : "Start block id",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "endblock",
          "in" : "query",
          "description" : "End block id",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of transactions",
            "content" : {
              "application-json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionPaginatedCollectionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "PaginationLinksResponse" : {
        "required" : [ "first", "last", "next", "prev" ],
        "type" : "object",
        "properties" : {
          "first" : {
            "type" : "string",
            "description" : "First page",
            "example" : "https://testnet.explorer.noah.network/blocks?page=1"
          },
          "last" : {
            "type" : "string",
            "description" : "Last page",
            "example" : "https://testnet.explorer.noah.network/blocks?page=10"
          },
          "prev" : {
            "description" : "Previous page",
            "nullable" : true,
            "example" : "https://testnet.explorer.noah.network/blocks?page=1"
          },
          "next" : {
            "description" : "Next page",
            "nullable" : true,
            "example" : "https://testnet.explorer.noah.network/blocks?page=2"
          }
        }
      },
      "PaginationMetaResponse" : {
        "required" : [ "current_page", "last_page", "path", "per_page", "total" ],
        "type" : "object",
        "properties" : {
          "current_page" : {
            "type" : "integer",
            "description" : "Current page number",
            "example" : 2
          },
          "last_page" : {
            "type" : "integer",
            "description" : "Last page number",
            "example" : 10
          },
          "path" : {
            "type" : "string",
            "description" : "Request path",
            "example" : "https://testnet.explorer.noah.network"
          },
          "per_page" : {
            "type" : "integer",
            "description" : "Elements per page",
            "example" : 50
          },
          "total" : {
            "type" : "integer",
            "description" : "Total number of elements",
            "example" : 16434
          }
        }
      },
...
      "TransactionResponse" : {
        "required" : [ "block", "data", "fee", "from", "hash", "nonce", "payload", "timestamp", "txn", "type" ],
        "properties" : {
          "txn" : {
            "type" : "integer",
            "description" : "Transaction number",
            "example" : 17653
          },
          "hash" : {
            "type" : "string",
            "description" : "Transaction hash",
            "example" : "Mtc923c18c1d842dcba262c63df22003d5ee4e683dc2927acf9580f083ad9e193e"
          },
          "nonce" : {
            "type" : "integer",
            "description" : "Transaction nonce",
            "example" : 113
          },
          "block" : {
            "type" : "integer",
            "description" : "Block id",
            "example" : 119
          },
          "timestamp" : {
            "type" : "string",
            "description" : "Transaction timestamp",
            "example" : "2019-02-19T00:00:00+03:00"
          },
          "fee" : {
            "type" : "string",
            "description" : "Transaction fee",
            "example" : "10.000000000000000000"
          },
          "type" : {
            "type" : "integer",
            "description" : "Transaction type",
            "example" : 1
          },
          "payload" : {
            "type" : "string",
            "description" : "Transaction payload",
            "example" : "Super tx from developers"
          },
          "from" : {
            "type" : "string",
            "description" : "Transaction sender address",
            "example" : "Noahx184ac726059e43643e67290666f7b3195093f870"
          },
          "data" : {
            "$ref" : "#/components/schemas/TxSendCoinData"
          }
        }
      },
...
      "TransactionPaginatedCollectionResponse" : {
        "required" : [ "data", "links", "meta" ],
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "description" : "Transactions",
            "items" : {
              "$ref" : "#/components/schemas/TransactionResponse"
            }
          },
          "links" : {
            "$ref" : "#/components/schemas/PaginationLinksResponse"
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMetaResponse"
          }
        }
      },
...
      "ValidatorResponse" : {
        "required" : [ "data" ],
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ValidatorResponse_data"
          }
        }
      },
      "StakeResponse" : {
        "required" : [ "address", "coin", "value" ],
        "properties" : {
          "coin" : {
            "type" : "string",
            "description" : "Coin",
            "example" : "MNT"
          },
          "address" : {
            "type" : "string",
            "description" : "Noah address",
            "example" : "Noahxe449e9fd85fd553fa45d245b62e30253bce29cc7"
          },
          "value" : {
            "type" : "string",
            "description" : "Amount",
            "example" : "38602.071957527232491487"
          }
        }
      },
...
      "TxSendCoinData" : {
        "properties" : {
          "coin" : {
            "type" : "string",
            "description" : "Send coin",
            "example" : "MNT"
          },
          "to" : {
            "type" : "string",
            "description" : "Receiver address",
            "example" : "Noahxee81347211c72524338f9680072af90744333146"
          },
          "value" : {
            "type" : "string",
            "description" : "Send coin amount",
            "example" : "12.357"
          }
        }
      },
...
      "inline_response_404_1_error" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "example" : 404
          },
          "message" : {
            "type" : "string",
            "example" : "Validator not found."
          }
        }
      },
      "inline_response_404_1" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/inline_response_404_1_error"
          }
        }
      },
      "ValidatorResponse_data" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "description" : "Validator status. 1 - not ready, 2 - ready",
            "example" : 2
          },
          "stake" : {
            "type" : "string",
            "description" : "Validator stake",
            "example" : "2456.78543535"
          },
          "part" : {
            "type" : "string",
            "description" : "Validator stake part of total stake",
            "example" : "0.33"
          },
          "delegator_count" : {
            "type" : "integer",
            "description" : "Delegators count",
            "example" : 12
          },
          "delegator_list" : {
            "type" : "array",
            "description" : "List of delegators",
            "items" : {
              "$ref" : "#/components/schemas/StakeResponse"
            }
          }
        },
        "description" : "Validator data"
      },
...
    }
  }
}